requirements.txt
streamlit==0.71.0
pandas==1.1.3
numpy==1.19.2
plotly-express==0.4.1
